requirements.txt
asgiref==3.3.1
astroid==2.4.1
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.6
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
Pillow==7.2.0
protobuf==3.15.5
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pylint==2.5.2
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
